requirements.txt
# Core Django + REST + Channels (WebSocket)
Django>=5.1,<6.0
djangorestframework>=3.14,<3.17
channels>=4.0,<5.0
channels_redis>=4.0,<5.0

# Authentication & JWT
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=65.0,<66.0   # अगर social login इस्तेमाल कर रहे हो, वरना हटा दो
gunicorn>=22.0,<23.0
# Admin UI enhancements
#jazzmin>=3.0,<4.0

stripe>=10.0,<11.0 


# Security & CORS
django-cors-headers>=4.0,<5.0
django-csp>=4.0,<5.0
django-ratelimit>=4.0,<5.0

# Filtering & utils
django-filter>=24.0,<25.0
django-simple-captcha 

# Database (PostgreSQL) - Railway पर binary इस्तेमाल करो ताकि compile न हो
psycopg2-binary>=2.9,<3.0

# Static files in production
whitenoise>=6.0,<7.0

# Celery (अगर tasks/background jobs हैं)
django-celery-beat>=2.5,<3.0
celery>=5.3,<6.0   # redis broker के साथ

# Environment variables (recommended)
python-decouple>=3.8,<4.0   # .env फाइल से variables पढ़ने के लिए

# Development / optional (local में रख सकते हो, production में नहीं जरूरी)
# debug-toolbar  # DEBUG=True पर
# black  # formatting
# flake8  # linting
